+++ src/helpers/OptionsHelper.ts
@@ -49,6 +49,7 @@ export class OptionsHelper {
         '--confirm': Boolean,
         '--skipExisting': Boolean,
         '--continueOnError': Boolean,
+        '--retryWhenFailed': Boolean,
         '-a': '--auth',
         '-f': '--folder',
         '-u': '--url'
@@ -77,6 +78,7 @@ export class OptionsHelper {
       skipPrecheck: args["--skipPrecheck"] as any || options["skipPrecheck"] || false,
       skipExistingPages: args["--skipExistingPages"] as any || options["skipExistingPages"] || false,
       continueOnError: args["--continueOnError"] as any || options["continueOnError"] || false,
+      retryWhenFailed: args["--retryWhenFailed"] as any || options["retryWhenFailed"] || false,
       menu: options["menu"] || null,
       debug: (process.env.DEBUG && process.env.DEBUG === "true") || args["--debug"] || false,
       cleanStart: args["--cleanStart"] || false,